# PERF: MeanImageFilter faster, using BufferedImageNeighborhoodPixelAccessPolicy