Improved the performance `MeanImageFilter`, using `ShapedImageNeighborhoodRange` with a newly added pixel access policy, `BufferedImageNeighborhoodPixelAccessPolicy`. Tested performance by:

    #include <itkMeanImageFilter.h>
    #include <itkImage.h>
    #include <chrono>
    #include <iostream>

    int main()
    {
      using ImageType = itk::Image<char, 3>;
      const auto image = ImageType::New();
      const auto size = ImageType::SizeType{ 1024, 1024, 16 };
      image->SetRegions(size);
      image->Allocate(true);

      const auto filter = itk::MeanImageFilter<ImageType, ImageType>::New();

      for (int i{}; i < 8; ++i)
      {
        filter->SetInput(nullptr);
        filter->SetInput(image);

        using namespace std::chrono;
        const auto timePointBeforeUpdate = high_resolution_clock::now();

        filter->Update();

        const auto timePointAfterUpdate = high_resolution_clock::now();

        const auto durationSeconds =
          duration_cast<duration<double>>(timePointAfterUpdate - timePointBeforeUpdate);
        std::cout
          << " Duration: " << durationSeconds.count() << " second(s)"
          << std::endl;
      }
    }