@dzenanz 2019-10-28
Does median filter have similar implementation pattern? Can this approach be applied to it?