@N-Dekker 2019-10-29
Thank you both for your comments!

@dzenanz 
> Does median filter have similar implementation pattern? Can this approach be applied to it?

Yes it does 😄 I just did `MeanImageFilter` with this PR, because it's slightly easier than `MedianImageFilter`. I do think that if this approach also appears applicable on other places, some extra utility could be helpful to limit the amount of copy-and-paste boilerplate code.

@blowekamp 
> As for the PR, are there any asserts of validations of the buffered region assumptions? or will be just get "undefined" behavior when the assumptions are not met?

An attempt to access a location outside the buffered region, using the proposed new pixel access policy, will produce undefined behavior, indeed. The calculation of the index value into the image buffer does not do bounds checking:
https://github.com/InsightSoftwareConsortium/ITK/blob/582b4c9792839098dbac3740d8bb8f5f792e0d95/Modules/Core/Common/include/itkBufferedImageNeighborhoodPixelAccessPolicy.h#L78

That's really a fundamental property of this new pixel access policy. It should be as fast as possible. Within `MeanImageFilter`, it is essential that it is _only_ being used with the first region produces by `ImageBoundaryFacesCalculator`, which is the center (non-boundary) region.

Some more profiling results from VS2017 Release:

| image size | duration before PR |  duration after PR |
| :--- | :--- | :--- |
| 256x256x256 | 0.26 sec | 0.15 sec |
| 16x1024x1024 | 0.49 sec | 0.16 sec |
| 1024x1024x16 | 1.4 (!!!) sec | 0.2 sec |

Could you please check if you can observe similar performance improvements from this PR?