@blowekamp 2019-10-29
I would suggest looking at the ConvolutionImageFilter next. There is easily room for 10x-100x performance improvement there.