@dzenanz 2019-10-29
`MeanImageFilter` being slightly easier than `MedianImageFilter` was my guess - I am glad to hear I was right 😄 

> But the checks for these assumptions (at least in debug mode)

`itkAssertInDebugAndIgnoreInReleaseMacro` or a regular C++ `assert` should work well. Of course, it will make debug mode somewhat slower, but it will be much easier to spot an error when it is introduced.

With included test, before: 0.3 sec, with this PR: 0.15 sec. 2x speedup, nice!