@N-Dekker 2019-10-29
Thanks for your comments @blowekamp 

> There are faster implementations of the Mean and Median filters in other implementations in ITK.

Which  faster implementation of the Mean filter do you mean?

Is the current (v5.1b01) `itk::MeanImageFilter` still being used, or is it outdated? It looks like my PR would make it twice as fast. Do you think that's worth the effort?