@blowekamp 2019-10-29
> Which faster implementation of the Mean filter do you mean?

The BoxMeanImageFilter performs better for higher dimension images and larger radius as it uses a moving accumulation. This implementation likely performs better for 2-D and smaller radius.

> Is the current (v5.1b01) `itk::MeanImageFilter` still being used, or is it outdated? It looks like my PR would make it twice as fast. Do you think that's worth the effort?

ITK is not about THE implementation for an algorithm. Many algorithms have multiple implementations which perform better in different situations. Others are frequent exemplar of how to do common operations in a generalized templated way. This filter I consider an exemplar filter for doing neighborhood operations, as well as giving good performance for small radius. It is most certainly worth the effort and is a great contribution.

Also do you know if these iterators work well with VectorImages? This filter did not work with them before.