@N-Dekker 2019-10-29
@dzenanz 
> `itkAssertInDebugAndIgnoreInReleaseMacro` or a regular C++ `assert` should work well.

Just added  an `assert`, please check:
https://github.com/InsightSoftwareConsortium/ITK/blob/16c31cad9fdf628b2cdcbde0761c39e8e55d7a2c/Modules/Core/Common/include/itkBufferedImageNeighborhoodPixelAccessPolicy.h#L85