@N-Dekker 2019-10-30
> Looks good, but CI has qualms with it.

Thanks @dzenanz I see at https://open.cdash.org/viewBuildError.php?type=1&buildid=6181856 
> itkBufferedImageNeighborhoodPixelAccessPolicy.h:76:50: warning: unused parameter 'imageSize' [-Wunused-parameter]

Even though I did `itkNotUsed(imageSize)`! I guess I have _not used_ `itkNotUsed` correctly  😺  Instead, I'll force-push an amend to just do a silly `void`-cast, `(void)imageSize;`, to silence the warning... OK?