@dzenanz 2019-10-30
What happens if you use `itkAssertInDebugAndIgnoreInReleaseMacro` instead of `assert`? If that doesn't help, you could apply this trick:
https://github.com/InsightSoftwareConsortium/ITK/blob/49671655c88612fabeb5f0784d3f40123100f166/Modules/Core/Common/include/itkImageSource.hxx#L247-L250