@N-Dekker 2019-10-30
@blowekamp 
> The BoxMeanImageFilter performs better for higher dimension images and larger radius as it uses a moving accumulation. This implementation likely performs better for 2-D and smaller radius.

Indeed: I just observed that with the "default" radius value (1), `MeanImageFilter` (with my proposed PR) appears significantly faster than `BoxMeanImageFilter`, even for 3-D... For a larger radius, I get mixed results: sometimes one is faster, sometimes the other.

The output of `BoxMeanImageFilter` is also slightly different than `MeanImageFilter`. It appears that `BoxMeanImageFilter` produces different values for border pixels. Is that a known issue?

> Also do you know if these iterators work well with VectorImages? This filter did not work with them before.

The iterator ranges from `Experimental` should work well with `VectorImage`. Looks like the incompatibility is caused by `sum = NumericTraits<InputRealType>::ZeroValue()`, in the filter itself: https://github.com/InsightSoftwareConsortium/ITK/blob/v5.1b01/Modules/Filtering/Smoothing/include/itkMeanImageFilter.hxx#L73

With this PR, I did not want to change that logic, because this PR is intended to just improve the performance of existing use cases. If you want to add `VectorImage` support to this filter, can you please make it a separate PR? Of course, I would like it if my PR (#1365) could be handled first  😄