@N-Dekker 2019-10-31
Succeeded on ITK.Linux, ITK.Windows, and ITK.macOS 🎉 

The ITK.Linux.Python failures at the dashboard seem unrelated to this PR: 

> AttributeError: module 'ITKMathematicalMorphology' has no attribute 'swig'
itkTestDriver: Process exited with return value: 1

https://open.cdash.org/testDetails.php?test=811772611&build=6183942