@dzenanz 2019-11-06
The assertion also gets triggered [here](https://open.cdash.org/testDetails.php?test=819819909&build=6195188), by a test you wrote @N-Dekker. So much about relying on people being careful, or however you phrased it Niels 😄