@N-Dekker 2019-11-07
> The assertion also gets triggered [here](https://open.cdash.org/testDetails.php?test=819819909&build=6195188), by a test you wrote @N-Dekker.

Thanks @dzenanz. I think it's related to issue #1397 (Valgrind Defects related to the MedianImageFilter), which should be fixed by pull request #1398 (BUG: Fix IndexRange for itk::Size that has one or more zero values)